IP名单新增IPv6和所有IP两种类型

+++ internal/utils/version.go
@@ -14,5 +14,5 @@ func VersionToLong(version string) uint32 {
 	} else if countDots == 0 {
 		version += ".0.0.0"
 	}
-	return IP2Long(version)
+	return uint32(IP2Long(version))
 }
